+++ editor/editor-mobile-bridge/src/__tests__/integration-webview/_mocks/types.ts
@@ -1,0 +1,8 @@
+import { MediaFile } from '@atlaskit/media-client';
+
+export type MockCollections = {
+  [key: string]: Array<MediaFile & { blob: Blob }>;
+};
+
+export type MockFileInputParams = Partial<MediaFile> & { dataUri?: string };
+export type MockFile = MediaFile & { blob?: Blob };