+++ editor/editor-mobile-bridge/webpack.test.config.js
@@ -1,0 +1,36 @@
+/* eslint import/no-extraneous-dependencies: 0*/
+const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const {
+  createBaseEditorMobileBridgeWebpackConfig,
+} = require('./build/webpack_base_config');
+
+module.exports = async function createWebpackConfig(_, args) {
+  return createBaseEditorMobileBridgeWebpackConfig(args, {
+    entry: {
+      editorTestSetup: path.join(
+        __dirname,
+        '/src/__tests__/integration-webview/_mocks/editorTestSetup.ts',
+      ),
+      editor: path.join(
+        __dirname,
+        '/src/__tests__/integration-webview/_mocks/editor-component.tsx',
+      ),
+    },
+    output: {
+      filename: '[name].js',
+      path: path.resolve(__dirname, 'dist/test/bundle'),
+    },
+    plugins: [
+      new HtmlWebpackPlugin({
+        template: path.join(__dirname, 'public/editor.html.ejs'),
+        chunks: ['editor', 'editorTestSetup'],
+        chunksSortMode: 'manual',
+        filename: 'editor.html',
+      }),
+    ],
+    devServer: {
+      contentBase: path.join(__dirname, 'dist/test/bundle'),
+    },
+  });
+};