+++ navigation/atlassian-navigation/src/components/Notifications/types.ts
@@ -5,7 +5,7 @@ import { BaseIconButtonProps } from '../IconButton/types';
 export type NotificationsProps = BaseIconButtonProps & {
   /**
    * Component to be used for the badge.
-   * Generally you'll want to use `NotificationIndicator` from [`@atlaskit/notification-indicator`](/packages/home/notification-indicator).
+   * Generally you'll want to use `NotificationIndicator` from [`@atlaskit/notification-indicator`](/packages/notifications/notification-indicator).
    */
   badge: ComponentType<{}>;
 };